requirements.txt
pillow==10.4.0
pillow_heif==0.18.0
